Applying kanop's PATCH "airbase-ng non-ASCII ESSID handling" from Ticket #1047
I've posted a report in the forums about airbase-ng not handling non-ASCII ESSID,
and Mister X committed a fix that disables the ASCII check on the ESSID (changeset [2208]).
Unfortunately, there are 2 more checks that cause airbase-ng to ignore non-ASCII probes,
or not handle them properly (not showing the associated ESSID).

This patch disables these checks.